move methods from calc to where they belong, delete calc module
