+++ nobubo/disassembly.py
@@ -18,17 +18,21 @@
 """
 Contains functions for various output layouts.
 """
+import math
 import pathlib
 from typing import List, Tuple, Optional
 
 import pikepdf
 
+import nobubo.assembly
 import nobubo.core
-from nobubo import core, calc, errors
+from nobubo import core, errors
+from nobubo.assembly import Layout, PageSize
+from nobubo.core import Factor
 
 
 def create_output_files(temp_collage_paths: List[pathlib.Path],
-                        input_properties: core.InputProperties,
+                        input_properties: nobubo.assembly.InputProperties,
                         output_properties: core.OutputProperties) -> None:
     for counter, collage_path in enumerate(temp_collage_paths):
         try:
@@ -36,7 +40,7 @@ def create_output_files(temp_collage_paths: List[pathlib.Path],
         except OSError as e:
             raise errors.UsageError("Could not open collage file for disassembly:"
                                     f"\n{e}.")
-        new_outputpath = calc.new_outputpath(output_properties.output_path, counter)
+        new_outputpath = generate_new_outputpath(output_properties.output_path, counter)
         print("\nChopping up the collage...")
         chopped_up_files = _create_output_files(collage, input_properties.pagesize,
                                                 input_properties.layout[counter],
@@ -58,7 +62,7 @@ def write_chops(collage: pikepdf.Pdf, output_path: pathlib.Path) -> None:
 def write_collage(temp_collage_paths: List[pathlib.Path],
                   output_properties: core.OutputProperties) -> None:
     for counter, collage_path in enumerate(temp_collage_paths):
-        new_outputpath = calc.new_outputpath(output_properties.output_path, counter)
+        new_outputpath = generate_new_outputpath(output_properties.output_path, counter)
         try:
             temp_collage = pikepdf.Pdf.open(collage_path)
             temp_collage.save(new_outputpath)
@@ -69,8 +73,8 @@ def write_collage(temp_collage_paths: List[pathlib.Path],
 
 
 def _create_output_files(collage: pikepdf.Pdf,
-                         pagesize: core.PageSize,
-                         current_layout: core.Layout,
+                         pagesize: nobubo.assembly.PageSize,
+                         current_layout: nobubo.assembly.Layout,
                          output_layout: Optional[List[int]]) -> pikepdf.Pdf:
     """
     Chops up the collage that consists of all the pattern pages to individual pages
@@ -81,7 +85,7 @@ def _create_output_files(collage: pikepdf.Pdf,
     :return: The pdf with several pages, ready to write to disk.
     """
     assert output_layout is not None
-    n_up_factor = calc.nup_factors(pagesize, output_layout)
+    n_up_factor = nup_factors(pagesize, output_layout)
     # only two points are needed to be cropped, lower left (x, y) and upper right (x, y)
     lowerleft_factor = nobubo.core.Factor(x=0, y=0)
     upperright_factor = nobubo.core.Factor(x=1, y=1)
@@ -90,7 +94,7 @@ def _create_output_files(collage: pikepdf.Pdf,
     output.copy_foreign(collage.Root)
     # Root must be copied too, not only the page:
     # thanks to https://github.com/cfcurtis/sewingutils
-    for i in range(0, calc.pages_needed(current_layout, n_up_factor)):
+    for i in range(0, pages_needed(current_layout, n_up_factor)):
         page = output.copy_foreign(collage.pages[0])
 
         lowerleft: core.Point = _calculate_lowerleft_point(lowerleft_factor,
@@ -121,15 +125,15 @@ def _calculate_colsrows_left(layout_element: int, factor: int, nup_factor: int) 
 
 def _calculate_lowerleft_point(lowerleft_factor: nobubo.core.Factor,
                                n_up_factor: nobubo.core.Factor,
-                               pagesize: core.PageSize) -> core.Point:
+                               pagesize: nobubo.assembly.PageSize) -> core.Point:
     return core.Point(x=lowerleft_factor.x * n_up_factor.x * pagesize.width,
                       y=lowerleft_factor.y * n_up_factor.y * pagesize.height)
 
 
 def _calculate_upperright_point(upperright_factor: nobubo.core.Factor,
                                 n_up_factor: nobubo.core.Factor,
-                                current_layout: core.Layout,
-                                pagesize: core.PageSize) -> core.Point:
+                                current_layout: nobubo.assembly.Layout,
+                                pagesize: nobubo.assembly.PageSize) -> core.Point:
     upperright = core.Point(x=0, y=0)
     # Manage ROWS: apply transformation to upper right, y-value
     rowsleft = _calculate_colsrows_left(current_layout.rows,
@@ -182,3 +186,38 @@ def _advance_vertically(lowerleft_factor: nobubo.core.Factor,
     upperright_factor.x = 1
     upperright_factor.y += 1
     return lowerleft_factor, upperright_factor
+
+
+def pages_needed(layout: Layout, n_up_factor: Factor) -> int:
+    x = layout.columns / n_up_factor.x
+    y = layout.rows / n_up_factor.y
+    return math.ceil(x) * math.ceil(y)
+
+
+def to_userspaceunits(width_height: List[int]) -> PageSize:
+    """
+    Converts a page's physical width and height from millimeters to
+    default user space unit, which is defined in the pdf standard as 1/72 inch.
+
+    :param width_height: Width and height of the physical page in millimeters (mm),
+    on which the pattern will be printed.
+    :return: Width and height of the physical page in default user space units.
+    """
+    # 1 mm = 5/127 inches = 0.03937 inches;  1/72 inch = 0.013888889
+    # conversion factor = 5/127 / 1/72 = 360/127 = 2.834645669
+    conversion_factor = 2.834645669
+
+    return PageSize(width=(round(width_height[0] * conversion_factor, 3)),
+                                    height=(round(width_height[1] * conversion_factor, 3)))
+
+
+def nup_factors(pagesize: PageSize, output_layout: List[int]) -> Factor:
+    output_papersize = to_userspaceunits(output_layout)
+    x_factor = int(output_papersize.width // pagesize.width)
+    y_factor = int(output_papersize.height // pagesize.height)
+    return Factor(x=x_factor, y=y_factor)
+
+
+def generate_new_outputpath(output_path: pathlib.Path, page_count: int) -> pathlib.Path:
+    new_filename = f"{output_path.stem}_{page_count + 1}{output_path.suffix}"
+    return output_path.parent / new_filename
